+++ extensions/positron-assistant/src/providers/azure/azureProvider.ts
@@ -10,10 +10,41 @@ import { ModelProvider } from '../base/modelProvider';
 import { ModelConfig } from '../../config';
 
 /**
- * Azure OpenAI model provider implementation.
- * Supports Azure OpenAI Service models.
+ * Azure OpenAI Service model provider implementation.
+ *
+ * This provider integrates with Azure OpenAI Service, which provides enterprise-grade
+ * access to OpenAI models with:
+ * - Azure's compliance and security features
+ * - Regional deployment options
+ * - Private endpoint support
+ * - All GPT models (GPT-4o, GPT-4 Turbo, etc.)
+ *
+ * **Configuration:**
+ * - Provider ID: `azure`
+ * - Required: API key and resource name from Azure Portal
+ * - Optional: Model selection, tool calling toggle
+ * - Default Model: GPT-4o
+ *
+ * @example
+ * ```typescript
+ * const config: ModelConfig = {
+ *   id: 'azure-gpt4o',
+ *   name: 'GPT 4o',
+ *   provider: 'azure',
+ *   apiKey: 'your-azure-key',
+ *   resourceName: 'your-resource-name',
+ *   model: 'gpt-4o'
+ * };
+ * const provider = new AzureLanguageModel(config, context);
+ * ```
+ *
+ * @see {@link ModelProvider} for base class documentation
+ * @see https://azure.microsoft.com/en-us/products/ai-services/openai-service for Azure OpenAI documentation
  */
 export class AzureLanguageModel extends ModelProvider implements positron.ai.LanguageModelChatProvider {
+	/**
+	 * The Azure OpenAI provider instance from Vercel AI SDK.
+	 */
 	protected declare aiProvider: AzureOpenAIProvider;
 
 	static source: positron.ai.LanguageModelSource = {
@@ -33,6 +64,7 @@ export class AzureLanguageModel extends ModelProvider implements positron.ai.Lan
 
 	constructor(_config: ModelConfig, _context?: vscode.ExtensionContext) {
 		super(_config, _context);
+		this.initializeLogger();
 		this.initializeProvider();
 	}
