+++ extensions/positron-assistant/src/providers/google/googleProvider.ts
@@ -10,10 +10,44 @@ import { ModelProvider } from '../base/modelProvider';
 import { ModelConfig, SecretStorage } from '../../config';
 
 /**
- * Google Generative AI (Gemini) model provider implementation.
- * Supports Google's Gemini models via the Generative AI API.
+ * Google Gemini model provider implementation.
+ *
+ * This provider integrates Google's Gemini models using the Vercel AI SDK's
+ * Google Generative AI adapter. Supports:
+ * - Gemini 2.0 Flash (latest)
+ * - Gemini 1.5 Pro
+ * - Gemini 1.5 Flash
+ * - Vision capabilities
+ * - Tool/function calling
+ * - Streaming responses
+ *
+ * **Configuration:**
+ * - Provider ID: `google`
+ * - Display Name: `Gemini Code Assist`
+ * - Required: API key from Google AI Studio
+ * - Optional: Custom base URL, model selection
+ * - Supports: Tool calling and completions
+ *
+ * @example
+ * ```typescript
+ * const config: ModelConfig = {
+ *   id: 'gemini-2-flash',
+ *   name: 'Gemini 2.0 Flash',
+ *   provider: 'google',
+ *   apiKey: 'your-api-key',
+ *   model: 'gemini-2.0-flash-exp',
+ *   baseUrl: 'https://generativelanguage.googleapis.com/v1beta'
+ * };
+ * const provider = new GoogleLanguageModel(config, context, storage);
+ * ```
+ *
+ * @see {@link ModelProvider} for base class documentation
+ * @see https://ai.google.dev/ for Google Generative AI documentation
  */
 export class GoogleLanguageModel extends ModelProvider implements positron.ai.LanguageModelChatProvider {
+	/**
+	 * The Google Generative AI provider instance from Vercel AI SDK.
+	 */
 	protected declare aiProvider: GoogleGenerativeAIProvider;
 
 	static source: positron.ai.LanguageModelSource = {
@@ -35,6 +69,7 @@ export class GoogleLanguageModel extends ModelProvider implements positron.ai.La
 
 	constructor(_config: ModelConfig, _context?: vscode.ExtensionContext, _storage?: SecretStorage) {
 		super(_config, _context, _storage);
+		this.initializeLogger();
 		this.initializeProvider();
 	}
